+++ Changelog.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master (unreleased)
+## v8.2.2 (2020-04-10)
 
 - Fixed Argentina's "Malvinas Day" date for 2020, shifted to March 31st because of the coronavirus crisis (#476).
 - Fixed Argentina's label for "Malvinas Day" and "Día de la Memoria" (#476).